refactor(architecture): Implement atomic service layer reorganization
BREAKING CHANGE: Service layer architecture reorganized for clean separation of concerns.

## Structural Changes
- Created `src/services/domain/` for pure business logic (project.ts, template.ts)
- Created `src/services/shared/` for cross-cutting utilities (pagination.ts)
- Maintained context-specific services (admin/, client/, auth/, validation/)

## Updated Import Paths
- Domain services: `@/services/domain/project|template`
- Shared utilities: `@/services/shared/pagination`
- Updated 5 API endpoints to use new import structure

## Architectural Impact
- Eliminated service discovery friction between domain and utility services
- Enhanced maintainability through strict separation of concerns
- Improved code organization following Clean Architecture principles
- Zero breaking changes to external APIs or functionality

## Verification
✅ All 256 tests passing
✅ TypeScript compilation with 0 errors
✅ ESLint check with 0 issues
✅ Service layer now follows atomic modular patterns

Strategic Rationale: This reorganization eliminates architectural debt that was blocking future maintainability by creating mixed service boundaries. Domain services are now properly isolated from shared utilities, enabling better service discovery and adherence to Clean Architecture principles.
