+++ docs/task.md
@@ -41,6 +41,7 @@
 - [x] **MEDIUM**: Enforce consistent error handling using `handleApiError()` utility across all API endpoints - ✅ 61 endpoints compliant
 - [x] **MODULARITY**: **HIGH IMPACT** - Extract dashboard inline business logic into client services - eliminated 150+ lines of duplicate code
 - [x] **MODULARITY**: **HIGH IMPACT** - Create validation service layer with domain-specific validators - eliminated 200+ lines of duplicate validation code
+- [x] **MODULARITY**: **ARCHITECTURAL** - Service layer reorganization into atomic structure - created `src/services/domain/` and `src/services/shared/` directories for clean separation of concerns, eliminated architectural friction between domain and utility services
 - [ ] **MEDIUM**: Add end-to-end integration tests for critical user flows (Registration → Order → Payment)
 - [x] **LOW**: Complete implementation of image optimization service in `src/lib/image-optimization.ts` - ✅ FULLY IMPLEMENTED
 - [ ] **LOW**: Reorganize domain services (`project.ts`, `template.ts`) into dedicated folder structure
