+++ src/services/domain/project.ts
@@ -4,7 +4,7 @@
  * Extracted from inline JavaScript in dashboard/projects.astro
  */
 
-import type { Project } from '../lib/types';
+import type { Project } from '../../lib/types';
 
 export interface ProjectStatus {
   label: string;
